conda_requirements.txt
pip
ipython
parso=0.7
flake8
black
mypy
numpy
scipy
matplotlib
requests
pandas
numba
beautifulsoup4
pytables
fastparquet
dask
python-blosc
tqdm